Refactor routes + api
- create a higher-level api by composing osm-p2p-db. Makes unit-testing easier
- remove dependency in routes on routes match object - allows use of different routers
