+++ handlers/changeset_close.js
@@ -3,13 +3,11 @@
  * http://wiki.openstreetmap.org/wiki/API_v0.6#Close:_PUT_.2Fapi.2F0.6.2Fchangeset.2F.23id.2Fclose
  */
 
-module.exports = function (closeChangeset) {
-  return function (req, res, osm, m, next) {
-    var parts = m.params.id.split(':')
-    var version = parts.length === 2 ? parts[1] : null
-    closeChangeset(parts[0], version, osm, function (err) {
-      if (err) return next(err)
-      res.end()
-    })
-  }
+module.exports = function (req, res, api, params, next) {
+  var parts = params.id.split(':')
+  var version = parts.length === 2 ? parts[1] : null
+  api.closeChangeset(parts[0], version, function (err) {
+    if (err) return next(err)
+    res.end()
+  })
 }
