+++ handlers/changeset_create.js
@@ -12,28 +12,26 @@ var xtend = require('xtend')
 var isValidContentType = require('../lib/valid_content_type.js')
 var xml2Obj = require('../transforms/xml_to_obj.js')
 
-module.exports = function (createChangeset) {
-  return function (req, res, osm, m, next) {
-    if (!isValidContentType(req)) {
-      return next(createError(400, 'unsupported content-type'))
-    }
-
-    var r = pumpify.obj(req, xmlNodes('changeset'), xml2Obj())
-    collect(r, function (err, ops) {
-      if (err) return next(createError(400, 'error parsing XML'))
-      if (!ops.length) return next(createError(400, 'no changeset elements found in XML'))
-
-      // If more than one changeset element is included in the PUT request
-      // tags are merged with later tags overwriting previous tags
-      var mergedChangeset = ops.reduce(function (p, v) {
-        p.tags = xtend(p.tags, v.tags)
-        return p
-      })
-      createChangeset(mergedChangeset, osm, function (err, id, node) {
-        if (err) return next(err)
-        res.setHeader('content-type', 'text/plain')
-        res.end(id + '\n')
-      })
-    })
+module.exports = function (req, res, api, params, next) {
+  if (!isValidContentType(req)) {
+    return next(createError(400, 'unsupported content-type'))
   }
+
+  var r = pumpify.obj(req, xmlNodes('changeset'), xml2Obj())
+  collect(r, function (err, ops) {
+    if (err) return next(createError(400, 'error parsing XML'))
+    if (!ops.length) return next(createError(400, 'no changeset elements found in XML'))
+
+    // If more than one changeset element is included in the PUT request
+    // tags are merged with later tags overwriting previous tags
+    var mergedChangeset = ops.reduce(function (p, v) {
+      p.tags = xtend(p.tags, v.tags)
+      return p
+    })
+    api.createChangeset(mergedChangeset, function (err, id, node) {
+      if (err) return next(err)
+      res.setHeader('content-type', 'text/plain')
+      res.end(id + '\n')
+    })
+  })
 }
