+++ lib/routes.js
@@ -1,52 +1,18 @@
 var router = require('routes')()
 
-var routes = [{
-  path: 'GET /api/(0.6)?/capabilities',
-  createHandler: require('../handlers/misc_capabilities.js')
-}, {
-  path: 'GET /api/0.6/map?*',
-  createHandler: require('../handlers/misc_map.js'),
-  operation: require('../operations/get_map.js')
-}, {
-  path: 'PUT /api/0.6/:type(changeset)/create',
-  createHandler: require('../handlers/changeset_create.js'),
-  operation: require('../operations/create_changeset.js')
-}, {
-  path: 'POST /api/0.6/changeset/:id/upload',
-  createHandler: require('../handlers/changeset_upload.js')
-}, {
-  path: 'GET /api/0.6/changeset/:id/download',
-  createHandler: require('../handlers/changeset_download.js'),
-  operation: require('../operations/get_changes.js')
-}, {
-  path: 'PUT /api/0.6/changeset/:id/close',
-  createHandler: require('../handlers/changeset_close.js'),
-  operation: require('../operations/close_changeset.js')
-}, {
-  path: 'PUT /api/0.6/:type(node|way|relation)/:id',
-  createHandler: require('../handlers/element_create.js'),
-  operation: require('../operations/create_element.js')
-}, {
-  path: 'GET /api/0.6/:type(node|way|relation)/:id/history',
-  createHandler: require('../handlers/element_history.js')
-}, {
-  path: 'GET /api/0.6/:type(node|way|relation)/:id/:version',
-  createHandler: require('../handlers/element_version.js'),
-  operation: require('../operations/get_element_version.js')
-}, {
-  path: 'GET /api/0.6/:type(nodes|ways|relations)\\?:ktype=:ids',
-  createHandler: require('../handlers/element_multi_fetch.js')
-}, {
-  path: 'GET /api/0.6/:type(node|way|relation)/:id',
-  createHandler: require('../handlers/element_read.js'),
-  operation: require('../operations/get_element.js')
-}, {
-  path: 'DELETE /api/0.6/:type(node|way|relation)/:id',
-  createHandler: require('../handlers/element_delete')
-}]
+function route (str, f) { router.addRoute(str, f) }
 
-routes.forEach(function (route) {
-  router.addRoute(route.path, route.createHandler(route.operation))
-})
+route('GET /api/(0.6)?/capabilities', require('../handlers/misc_capabilities.js'))
+route('GET /api/0.6/map?*', require('../handlers/misc_map.js'))
+route('PUT /api/0.6/:type(changeset)/create', require('../handlers/changeset_create.js'))
+route('POST /api/0.6/changeset/:id/upload', require('../handlers/changeset_upload.js'))
+route('GET /api/0.6/changeset/:id/download', require('../handlers/changeset_download.js'))
+route('PUT /api/0.6/changeset/:id/close', require('../handlers/changeset_close.js'))
+route('PUT /api/0.6/:type(node|way|relation)/:id', require('../handlers/element_create.js'))
+route('GET /api/0.6/:type(node|way|relation)/:id/history', require('../handlers/element_history.js'))
+route('GET /api/0.6/:type(node|way|relation)/:id/:version', require('../handlers/element_version.js'))
+route('GET /api/0.6/:type(nodes|ways|relations)\\?:ktype=:ids', require('../handlers/element_multi_fetch.js'))
+route('GET /api/0.6/:type(node|way|relation)/:id', require('../handlers/element_read.js'))
+route('DELETE /api/0.6/:type(node|way|relation)/:id', require('../handlers/element_delete'))
 
 module.exports = router
